requirements.txt
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
constantly==15.1.0
coverage==4.4.1
cryptography==1.8.1
cssselect==1.0.1
enum34==1.1.6
funcsigs==1.0.2
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
lxml==3.8.0
mock==2.0.0
nose==1.3.7
nose-watch==0.9.1
packaging==16.8
parsel==1.1.0
pathtools==0.1.2
pbr==3.0.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
python-dotenv==0.6.4
PyYAML==3.12
queuelib==1.4.2
requests==2.17.3
Scrapy==1.3.3
service-identity==16.0.0
six==1.10.0
Twisted==17.1.0
urllib3==1.21.1
w3lib==1.17.0
watchdog==0.8.3
zope.interface==4.4.0
